Make Card shareable
