+++ apps/ff-site/src/app/_styles/globals.css
@@ -219,4 +219,33 @@
   .slide-over-panel {
     @apply bg-brand-800;
   }
+
+  /* CARD */
+  .card {
+    @apply rounded-lg bg-brand-700/30;
+
+    &.card-border-lighter {
+      @apply border-brand-300;
+    }
+
+    &.card-border-light {
+      @apply border-brand-400;
+    }
+
+    &.card-border-base {
+      @apply border-brand-500;
+    }
+
+    &.card-border-dark {
+      @apply border-brand-600;
+    }
+  }
+
+  .card-image {
+    @apply rounded-lg px-1 pt-1;
+  }
+
+  .card-link {
+    @apply text-brand-300;
+  }
 }
