+++ packages/ui/package.json
@@ -17,6 +17,7 @@
     "./BreadCrumbs": "./src/BreadCrumbs.tsx",
     "./BreadCrumbsLayout": "./src/BreadCrumbsLayout.tsx",
     "./BreakpointDebugger": "./src/BreakpointDebugger.tsx",
+    "./Card": "./src/Card.tsx",
     "./DescriptionText": "./src/DescriptionText.tsx",
     "./DigestArticleHeader": "./src/DigestArticleHeader.tsx",
     "./ErrorMessage": "./src/ErrorMessage.tsx",
@@ -25,6 +26,7 @@
     "./Icon": "./src/Icon.tsx",
     "./IconButton": "./src/IconButton.tsx",
     "./MarkdownPage": "./src/MarkdownPage.tsx",
+    "./Meta": "./src/Meta.tsx",
     "./NoResultsMessage": "./src/NoResultsMessage.tsx",
     "./NoSearchResultsMessage": "./src/NoSearchResultsMessage.tsx",
     "./PageLayout": "./src/PageLayout.tsx",
